feat(modules/generic.ts): add `getStorageDir()` function
A preparation patch to implement `/warn`, `/unwarn` and `/warns`
commands. They need to remember the number of warnings that the users
got, use this storage directory to save that state. Also, introduce a
new environment variable `STORAGE_DIR`.

We will use a real database in the future. Irvan said, he planned to
use a PostgreSQL server.

Signed-off-by: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>