fix(modules/admin/warn.ts): Fix negative number of warns
When a user is never warned, but get unwarned, the number of warns
shown on Telegram will be negative. This is because @nr_warn is -1
and it returns @nr_warn directly without any negative guard if the
user doesn't have a user_file.json.

Fix this by guarding the negative value, and replace it with 0.

Link: https://t.me/deeznutsisbig/1362
Discovered-by: Irvan Malik <irvanmalik48@gmail.com>
Signed-off-by: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>